requirements_whatsapp.txt
# Dependencias principales
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio>=5.10.0
gevent>=24.0.0
gevent-websocket>=0.10.1

# WhatsApp Web
pywhatkit>=5.4
selenium>=4.15.0
pillow>=10.0.0

# QR y utilidades
qrcode[pil]>=7.4.0
requests>=2.31.0
python-dotenv>=1.0.0

# Base de datos y almacenamiento
redis>=5.0.0

# Logging y utilidades
colorama>=0.4.0
rich>=13.0.0

# Servidor web
gunicorn>=21.0.0
